+++ src/tests/intrinsic.cl
@@ -5,7 +5,7 @@ kernel void MyKernel()
 	uint3 dtid = (uint3)(get_global_id(0u), get_global_id(1u), get_global_id(2u));
 	barrier(CLK_LOCAL_MEM_FENCE | CLK_GLOBAL_MEM_FENCE);
 	barrier(CLK_LOCAL_MEM_FENCE | CLK_GLOBAL_MEM_FENCE);
-	int i = 0;
+	int i = (int)0;
 	int i1;
 	int2 i2;
 	int3 i3;
